requirements.txt
Flask==2.3.2
qrcode==7.4.2
cryptography==41.0.1
Pillow==9.5.0
